Extract the connection substrate from the gateway
